+++ packages/css-color-parser/test/wpt/index.mjs
@@ -15,4 +15,5 @@ import './color-invalid-hex.mjs';
 import './color-invalid-hsl.mjs';
 import './color-invalid-hwb.mjs';
 import './color-invalid-named-color.mjs';
+import './color-invalid-relative-color.mjs';
 import './color-invalid-rgb.mjs';